chore: refactor scripts to fix dual modules
